requirements.txt
absl-py==2.1.0
altair==5.4.1
astmonkey==0.3.6
astunparse==1.6.3
attrs==24.2.0
Automat==24.8.1
blinker==1.9.0
bytecode==0.16.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
code2flow==2.5.1
colorama==0.4.6
commonmark==0.9.1
constantly==23.10.4
construct==2.10.70
dot==0.3.0
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.68.0
h5py==3.12.1
huggingface-hub==0.26.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
jellyfish==0.11.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
libclang==18.1.1
mando==0.7.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
MutPy-Pynguin==0.7.1
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.14.2
networkx==2.8.8
numpy==2.1.3
opt_einsum==3.4.0
optree==0.13.1
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
protobuf==5.28.3
pyarrow==18.0.0
pyasn1==0.6.1
pydeck==0.9.1
pydot==1.4.2
pyflowchart==0.3.1
Pygments==2.18.0
pynguin==0.19.0
pyparsing==3.2.0
pysnmp==7.1.13
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
radon==6.0.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
simple-parsing==0.0.18
six==1.16.0
smmap==5.0.1
streamlit==1.40.1
streamlit-ace==0.1.1
streamlit-flow==0.1.0
streamlit-scroll-navigation==1.2.1
sympy==1.13.3
tenacity==9.0.0
tensor==0.3.6
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tokenizers==0.20.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.0
transformers==4.46.3
Twisted==24.10.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
zope.interface==7.1.1
